Update typescript rules to work in sandbox.
This is... hell-a-hacky. I should investigate getting a better setup for
node based binaries, but this works so I'm going to go with it for now.
I need to get unblocked on a personal project.

This is kind of a path forward though since it works with Bazel 0.2.2.

Next up will be making ts_library rules. I think that the plumbing is
there to handle it so it shouldn't be super difficult (famous last
words).